requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
playwright>=1.49.0
python-dotenv>=1.0.0
streamlit>=1.40.0
pandas>=2.2.0
httpx>=0.27.0
python-multipart>=0.0.9
